% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataFromLinks.R
\name{extractDataFromLinks}
\alias{extractDataFromLinks}
\title{A Webscrape Function}
\usage{
extractDataFromLinks(url, links)
}
\arguments{
\item{url}{url where table exists.}

\item{links}{a vector containing links}
}
\description{
This function allows you to extract data from series of links.
}
\examples{
extractDataFromLinks()
}
