% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLinksFromTable.R
\name{extractLinksFromTable}
\alias{extractLinksFromTable}
\title{A Webscrape Function}
\usage{
extractLinksFromTable(url, tableID)
}
\arguments{
\item{url}{url where table exists.}

\item{tableID}{id attribute of the table}
}
\description{
This function allows you to extract all the links from table.
}
\examples{
extractLinksFromTable()
}
