% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTableData.R
\name{extractTableData}
\alias{extractTableData}
\title{A Webscrape Function}
\usage{
extractTableData(url, tableID)
}
\arguments{
\item{url}{url where table exists.}

\item{tableID}{id attribute of the table}
}
\description{
This function allows you to extract all the data from table and returns dataframe
}
\examples{
extractTableData("https://www.drugshortagescanada.ca","drug-shortage-reports")
}
