% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElementAttr.R
\name{getElementAttr}
\alias{getElementAttr}
\title{A Webscrape Function}
\usage{
getElementAttr(html, cssSelector, Attr)
}
\arguments{
\item{html}{html code of the page}

\item{cssSelector}{css selector of the element}

\item{Attr}{attribute name whose value to be fetched}
}
\description{
This function allows you to extract attribute value from the element and returns vector.
}
\examples{
getElementAttr()
}
