% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportTable.R
\name{getReportTable}
\alias{getReportTable}
\title{A Webscrape Function}
\usage{
getReportTable(html, n)
}
\arguments{
\item{html}{html code of the page}

\item{n}{sequence of the table on the page}
}
\description{
This function gives to table in R based on position of table on the page.
}
\examples{
getReportTable()
}
