% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElementText.R
\name{getElementText}
\alias{getElementText}
\title{A Webscrape Function}
\usage{
getElementText(html, cssSelector)
}
\arguments{
\item{html}{html code of the page}

\item{cssSelector}{css selector of the element}
}
\description{
This function allows you to extract text from an element and returns vector
}
\examples{
html <- xml2::read_html("https://www.drugshortagescanada.ca")
getElementText(html,"table tbody tr td")
}
